# docs/openapi.yaml
openapi: 3.0.3
info:
  title: LEB2SCRAPPER API
  description: >
    HTTP contract for the LEB2SCRAPPER API. This file is a machine-readable mirror of
    docs/api-reference.md, which remains the canonical prose reference (migration
    rollout order, Supabase schema/migration SQL, device-binding and client-compatibility
    behavior). Every request/response body uses JSON unless stated otherwise.


    API versioning is expressed in the URL path, not in this document's `info.version`.
    `v1` is served by default; `ActivityController` additionally serves `v2`, which
    returns identical fields to `v1` but serializes `startDate`, `dueDate`, `createdAt`,
    `lastDueDateNotificationDate`, and `lastStatusChangeNotificationDate` as genuine UTC
    instants instead of naive GMT+7 wall-clock strings. Deprecated unversioned legacy
    alias routes (enabled while `ApiVersioning:LegacyRoutesEnabled=true`) execute the
    same `v1` actions and are omitted from this document, matching their exclusion from
    the generated Swagger documents.
  version: "1.0.0"
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:5015
    description: Local development
  - url: "https://<cloud-run-service-url>"
    description: Cloud Run (URL is per-deployment)
...
paths:
...
  /api/v1/Activity/{semesterId}/{classId}:
    get:
      tags: [Activity]
      operationId: getActivitiesByClassV1
      summary: List activities for one class (v1, legacy naive GMT+7 date shape)
      description: >
        Legacy shape — see `GET /api/v2/Activity/{semesterId}/{classId}` for the
        timezone-corrected version. Discovers classes for `semesterId` through the
        60-second structural cache and verifies `classId` belongs to that semester
        before retrieving activities; if the relationship is missing, the response is
        `404 RESOURCE_NOT_FOUND` and the activity lookup is not performed.
      security:
        - AccessKeyAuth: []
          Leb2BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SemesterIdPath"
        - $ref: "#/components/parameters/ClassIdPath"
        - $ref: "#/components/parameters/UserIdHeader"
        - $ref: "#/components/parameters/DeviceIdHeader"
        - $ref: "#/components/parameters/DeviceNameHeader"
        - $ref: "#/components/parameters/DevicePlatformHeader"
        - $ref: "#/components/parameters/DeviceOsVersionHeader"
        - $ref: "#/components/parameters/ClientVersionHeader"
      responses:
        "200":
          description: Activities for the class (empty array if none were found).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityV1"
        "400":
          description: >
            A route integer was less than one, or `X-LEB2-USER-ID` was missing,
            non-integer, or less than one.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
        "401":
          $ref: "#/components/responses/ActivityUnauthorized"
        "403":
          $ref: "#/components/responses/ActivityForbidden"
        "404":
          description: "`RESOURCE_NOT_FOUND` — the class does not belong to the supplied semester."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
        "502":
          $ref: "#/components/responses/BadGateway"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/v2/Activity/{semesterId}/{classId}:
    get:
      tags: [Activity]
      operationId: getActivitiesByClassV2
      summary: List activities for one class (v2, UTC date shape)
      description: >
        Identical to `GET /api/v1/Activity/{semesterId}/{classId}` in route
        parameters, headers, discovery behavior, and error responses. The only
        difference is that `startDate`, `dueDate`, `createdAt`,
        `lastDueDateNotificationDate`, and `lastStatusChangeNotificationDate`
        serialize as genuine UTC instants instead of naive GMT+7 wall-clock strings.
      security:
        - AccessKeyAuth: []
          Leb2BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SemesterIdPath"
        - $ref: "#/components/parameters/ClassIdPath"
        - $ref: "#/components/parameters/UserIdHeader"
        - $ref: "#/components/parameters/DeviceIdHeader"
        - $ref: "#/components/parameters/DeviceNameHeader"
        - $ref: "#/components/parameters/DevicePlatformHeader"
        - $ref: "#/components/parameters/DeviceOsVersionHeader"
        - $ref: "#/components/parameters/ClientVersionHeader"
      responses:
        "200":
          description: Activities for the class (empty array if none were found).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityV2"
        "400":
          description: >
            A route integer was less than one, or `X-LEB2-USER-ID` was missing,
            non-integer, or less than one.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
        "401":
          $ref: "#/components/responses/ActivityUnauthorized"
        "403":
          $ref: "#/components/responses/ActivityForbidden"
        "404":
          description: "`RESOURCE_NOT_FOUND` — the class does not belong to the supplied semester."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
        "502":
          $ref: "#/components/responses/BadGateway"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/v1/Activity/{semesterId}:
    get:
      tags: [Activity]
      operationId: getActivitiesBySemesterV1
      summary: List activities for every class in a semester (v1, legacy naive GMT+7 date shape)
      description: >
        Legacy shape — see `GET /api/v2/Activity/{semesterId}` for the
        timezone-corrected version. Discovers every class in the semester, retrieves
        their activities with maximum parallelism two, and returns one flat array.
        Classes are processed in class ID order; LEB2's activity order is preserved
        within each class. Fail-fast: an error is returned instead of a partial list
        if class discovery or any class activity request fails.
      security:
        - AccessKeyAuth: []
          Leb2BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SemesterIdPath"
        - $ref: "#/components/parameters/UserIdHeader"
        - $ref: "#/components/parameters/DeviceIdHeader"
        - $ref: "#/components/parameters/DeviceNameHeader"
        - $ref: "#/components/parameters/DevicePlatformHeader"
        - $ref: "#/components/parameters/DeviceOsVersionHeader"
        - $ref: "#/components/parameters/ClientVersionHeader"
      responses:
        "200":
          description: >
            Activities for the semester (empty array if the semester has no
            published classes or activities).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityV1"
        "400":
          description: "`semesterId` was less than one, or `X-LEB2-USER-ID` was missing, non-integer, or less than one."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
        "401":
          $ref: "#/components/responses/ActivityUnauthorized"
        "403":
          $ref: "#/components/responses/ActivityForbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
        "502":
          $ref: "#/components/responses/BadGateway"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/v2/Activity/{semesterId}:
    get:
      tags: [Activity]
      operationId: getActivitiesBySemesterV2
      summary: List activities for every class in a semester (v2, UTC date shape)
      description: >
        Identical to `GET /api/v1/Activity/{semesterId}` in route parameters, headers,
        discovery behavior, and error responses. The only difference is that
        `startDate`, `dueDate`, `createdAt`, `lastDueDateNotificationDate`, and
        `lastStatusChangeNotificationDate` serialize as genuine UTC instants instead
        of naive GMT+7 wall-clock strings.
      security:
        - AccessKeyAuth: []
          Leb2BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SemesterIdPath"
        - $ref: "#/components/parameters/UserIdHeader"
        - $ref: "#/components/parameters/DeviceIdHeader"
        - $ref: "#/components/parameters/DeviceNameHeader"
        - $ref: "#/components/parameters/DevicePlatformHeader"
        - $ref: "#/components/parameters/DeviceOsVersionHeader"
        - $ref: "#/components/parameters/ClientVersionHeader"
      responses:
        "200":
          description: >
            Activities for the semester (empty array if the semester has no
            published classes or activities).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityV2"
        "400":
          description: "`semesterId` was less than one, or `X-LEB2-USER-ID` was missing, non-integer, or less than one."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
        "401":
          $ref: "#/components/responses/ActivityUnauthorized"
        "403":
          $ref: "#/components/responses/ActivityForbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
        "502":
          $ref: "#/components/responses/BadGateway"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/v1/Activity/{semesterId}/snapshot:
    get:
      tags: [Activity]
      operationId: getSemesterSnapshotV1
      summary: Nested classes-and-activities snapshot for a semester (v1, legacy naive GMT+7 date shape)
      description: >
        Legacy shape — see `GET /api/v2/Activity/{semesterId}/snapshot` for the
        timezone-corrected version. Classes with no activities remain in the
        response. Fail-fast: never returns a successful partial snapshot.
      security:
        - AccessKeyAuth: []
          Leb2BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SemesterIdPath"
        - $ref: "#/components/parameters/UserIdHeader"
        - $ref: "#/components/parameters/DeviceIdHeader"
        - $ref: "#/components/parameters/DeviceNameHeader"
        - $ref: "#/components/parameters/DevicePlatformHeader"
        - $ref: "#/components/parameters/DeviceOsVersionHeader"
        - $ref: "#/components/parameters/ClientVersionHeader"
      responses:
        "200":
          description: "Nested snapshot (empty `classes` array if the semester has no published classes)."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SemesterSnapshotResponseV1"
        "400":
          description: "`semesterId` was less than one, or `X-LEB2-USER-ID` was missing, non-integer, or less than one."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
        "401":
          $ref: "#/components/responses/ActivityUnauthorized"
        "403":
          $ref: "#/components/responses/ActivityForbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
        "502":
          $ref: "#/components/responses/BadGateway"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/v2/Activity/{semesterId}/snapshot:
    get:
      tags: [Activity]
      operationId: getSemesterSnapshotV2
      summary: Nested classes-and-activities snapshot for a semester (v2, UTC date shape)
      description: >
        Identical to `GET /api/v1/Activity/{semesterId}/snapshot` in route
        parameters, headers, nesting behavior, and error responses. The only
        difference is that each nested activity's `startDate`, `dueDate`,
        `createdAt`, `lastDueDateNotificationDate`, and
        `lastStatusChangeNotificationDate` serialize as genuine UTC instants instead
        of naive GMT+7 wall-clock strings.
      security:
        - AccessKeyAuth: []
          Leb2BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SemesterIdPath"
        - $ref: "#/components/parameters/UserIdHeader"
        - $ref: "#/components/parameters/DeviceIdHeader"
        - $ref: "#/components/parameters/DeviceNameHeader"
        - $ref: "#/components/parameters/DevicePlatformHeader"
        - $ref: "#/components/parameters/DeviceOsVersionHeader"
        - $ref: "#/components/parameters/ClientVersionHeader"
      responses:
        "200":
          description: "Nested snapshot (empty `classes` array if the semester has no published classes)."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SemesterSnapshotResponseV2"
        "400":
          description: "`semesterId` was less than one, or `X-LEB2-USER-ID` was missing, non-integer, or less than one."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
        "401":
          $ref: "#/components/responses/ActivityUnauthorized"
        "403":
          $ref: "#/components/responses/ActivityForbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
        "502":
          $ref: "#/components/responses/BadGateway"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
components:
  securitySchemes:
    AccessKeyAuth:
      type: apiKey
      in: header
      name: access-key
      description: >
        Provisioned UUID from Supabase `keys.id`. A secret per-user API credential,
        not a JWT. Case-insensitive header name; exactly one UUID value. Never
        accepted in a URL, query string, or request body. `POST /api/v1/User/login`
        accepts a provisioned-but-unassigned key; every other protected route
        requires an assigned (activated) key.
    Leb2BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Opaque LEB2 session cookie
      description: >
        `Authorization: Bearer <leb2-session-cookie>`. The value is the opaque
        cookie string returned by `POST /api/v1/User/cookie`, not a JWT — it is
        never locally parsed or verified; actual validity is determined by LEB2's
        responses. Legacy clients may send the cookie without the `Bearer` prefix.
  parameters:
    SemesterIdPath:
      name: semesterId
      in: path
      required: true
      description: LEB2 semester ID.
      schema:
        type: integer
        minimum: 1
      example: 101
    ClassIdPath:
      name: classId
      in: path
      required: true
      description: LEB2 class ID used for the activity lookup.
      schema:
        type: integer
        minimum: 1
      example: 3001
    UserIdHeader:
      name: X-LEB2-USER-ID
      in: header
      required: true
      description: >
        Client-supplied assertion of the numeric LEB2 identity, kept for
        compatibility. Must match the LEB2 identity stored for the access-key
        owner; the opaque `Authorization` session value is never parsed to derive
        identity.
      schema:
        type: integer
        minimum: 1
      example: 2001
    DeviceIdHeader:
      name: X-Device-ID
      in: header
      required: false
      description: >
        Stable app-generated device identifier. HMAC-SHA256'd server-side with
        `DeviceBinding:HmacSecret` before storage; the raw value never enters
        PostgreSQL logs or response bodies. Optional while
        `DeviceBinding:EnforcementEnabled=false`, required once enforcement is
        enabled.
      schema:
        type: string
    DeviceNameHeader:
      name: X-Device-Name
      in: header
      required: false
      description: Display name for the device.
      schema:
        type: string
      example: Example phone
    DevicePlatformHeader:
      name: X-Device-Platform
      in: header
      required: false
      schema:
        type: string
      example: android
    DeviceOsVersionHeader:
      name: X-Device-OS-Version
      in: header
      required: false
      schema:
        type: string
      example: "14"
    ClientVersionHeader:
      name: X-Client-Version
      in: header
      required: false
      description: >
        Authoritative frontend application version (semantic version). Populates
        the device binding's `app_version` when device binding is persisted.
        Optional while `ClientCompatibility:EnforcementEnabled=false`; once
        enabled, missing/blank values receive `400 CLIENT_VERSION_REQUIRED` and
        malformed or repeated values receive `400 CLIENT_VERSION_INVALID`. A
        semantically older supported client receives
        `426 CLIENT_UPDATE_REQUIRED`.
      schema:
        type: string
      example: 0.5.2
  responses:
...
    ActivityUnauthorized:
      description: >
        `AUTHENTICATION_REQUIRED` (session header absent/empty), `ACCESS_KEY_REQUIRED`,
        `ACCESS_KEY_INVALID`, or `SESSION_EXPIRED` (LEB2 rejected the session —
        includes `WWW-Authenticate: Bearer`).
      headers:
        WWW-Authenticate:
          required: false
          schema:
            type: string
            example: Bearer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ActivityForbidden:
      description: >
        `ACCESS_KEY_NOT_ACTIVATED` (key not claimed through `/api/v1/User/login`),
        `ACCESS_KEY_IDENTITY_MISMATCH` (`X-LEB2-USER-ID` does not match the bound
        LEB2 identity), or `ACCESS_KEY_REAUTHENTICATION_REQUIRED` (stored LEB2
        identity not initialized).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    TooManyRequests:
      description: "`CLIENT_THROTTLE_ACTIVE` — this client has too many active or queued LEB2 requests."
      headers:
        Retry-After:
          required: true
          schema:
            type: integer
            minimum: 1
          description: Seconds until the client's request queue is expected to clear.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: "`UNEXPECTED_ERROR` — an unexpected server error occurred."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadGateway:
      description: "`LEB2_UNAVAILABLE` (LEB2 rejected or could not complete the request) or `SCRAPE_RESPONSE_CHANGED` (unexpected HTML/JSON structure)."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ServiceUnavailable:
      description: >
        `LEB2_UNAVAILABLE` (transient LEB2 or local browser-automation failure),
        `REQUEST_BACKOFF_ACTIVE` (this operation's backoff scope is paused after a
        recent upstream failure), or `ACCESS_KEY_STORE_UNAVAILABLE` (Supabase
        access-key validation is temporarily unavailable). `Retry-After` is present
        only for `REQUEST_BACKOFF_ACTIVE`.
      headers:
        Retry-After:
          required: false
          schema:
            type: integer
            minimum: 1
          description: Seconds until this backoff scope is expected to clear. Present only for `REQUEST_BACKOFF_ACTIVE`.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      description: Standard shape for operational, authentication, and upstream errors.
      required: [message, responseCode, details, timestamp, traceId]
      properties:
        message:
          type: string
          example: LEB2 is temporarily unavailable.
        responseCode:
          $ref: "#/components/schemas/ApiErrorCode"
        details:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time
          example: "2026-07-24T12:00:00Z"
        traceId:
          type: string
          nullable: true
          example: example-trace-id
    ValidationErrorResponse:
      type: object
      description: Shape for request-model, route-value, and header validation errors.
      required: [statusCode, message, responseCode, timestamp, traceId, validationErrors]
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          type: string
          example: Validation failed.
        responseCode:
          $ref: "#/components/schemas/ApiErrorCode"
        timestamp:
          type: string
          format: date-time
          example: "2026-07-24T12:00:00Z"
        traceId:
          type: string
          nullable: true
          example: example-trace-id
        validationErrors:
          type: object
          nullable: true
          description: Keys and messages depend on which input failed.
          additionalProperties:
            type: array
            items:
              type: string
          example:
            semesterId:
              - The field semesterId must be between 1 and 2147483647.
...
    ActivityV1:
      description: >
        Legacy shape (and all unversioned legacy-alias routes). `startDate`,
        `dueDate`, `createdAt`, `lastDueDateNotificationDate`, and
        `lastStatusChangeNotificationDate` are naive GMT+7 wall-clock strings with
        no offset — see "Why `/api/v2/Activity/...` exists" in
        docs/api-reference.md.
      allOf:
        - $ref: "#/components/schemas/ActivityCommon"
        - type: object
          required: [startDate, dueDate, createdAt, lastDueDateNotificationDate, lastStatusChangeNotificationDate]
          properties:
            startDate:
              type: string
              nullable: true
              pattern: '^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}$'
              description: Naive GMT+7 wall-clock, no offset.
              example: "2026-07-01T09:00:00"
            dueDate:
              type: string
              nullable: true
              pattern: '^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}$'
              description: Naive GMT+7 wall-clock, no offset.
              example: "2026-07-31T23:59:00"
            createdAt:
              type: string
              pattern: '^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}$'
              description: Naive GMT+7 wall-clock, no offset.
              example: "2026-06-30T12:00:00"
            lastDueDateNotificationDate:
              type: string
              nullable: true
              pattern: '^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}$'
              description: Naive GMT+7 wall-clock, no offset.
              example: null
            lastStatusChangeNotificationDate:
              type: string
              nullable: true
              pattern: '^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}$'
              description: Naive GMT+7 wall-clock, no offset.
              example: null
...
    SemesterSnapshotResponseV1:
      type: object
      required: [semesterId, classes]
      properties:
        semesterId:
          type: integer
          example: 101
        classes:
          type: array
          items:
            $ref: "#/components/schemas/SemesterSnapshotClassV1"
    SemesterSnapshotResponseV2:
      type: object
      required: [semesterId, classes]
      properties:
        semesterId:
          type: integer
          example: 101
        classes:
          type: array
          items:
            $ref: "#/components/schemas/SemesterSnapshotClassV2"
    SemesterSnapshotClassV1:
      type: object
      required: [id, name, activities]
      properties:
        id:
          type: integer
          example: 3001
        name:
          type: string
          example: Example Course
        activities:
          type: array
          items:
            $ref: "#/components/schemas/ActivityV1"
...